backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
openai
websockets
deepgram-sdk
elevenlabs==1.3.0
python-multipart
